Changed AddJSONObjectIntegerItem to Long

+++ src/main/java/org/jspringbot/keyword/json/AddJSONObjectIntegerItem.java
@@ -31,7 +31,7 @@ public class AddJSONObjectIntegerItem extends AbstractJSONKeyword {
 
     @Override
     public Object execute(Object[] params) throws Exception {
-            creator.addJSONObjectItem(String.valueOf(params[0]), Integer.valueOf(String.valueOf(params[1])));
+            creator.addJSONObjectItem(String.valueOf(params[0]), Long.valueOf(String.valueOf(params[1])));
 
         return null;
     }
